+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This file is used to list changes made in the last 3 major versions of the postg
 
 ## Unreleased
 
+- Fix attribute updates for users with dashes
+
 ## 8.1.0 - *2020-12-09*
 
 - Fix potential password exposure in logs